==> scripts/03-test_raw_data.R <==
#### Preamble ####
# Purpose: Tests the preliminary cleaned data sets before coding them for severity of harassment & type of account
# Author: Inessa De Angelis
# Date: 16 July 2023
# Contact: inessa.deangelis@mail.utoronto.ca 
# License: MIT
# Pre-requisites
  # 01-raw_data_cleaning.R

#### Workspace setup ####
library(tidyverse)
library(testthat)

#### Read in the preliminary cleaned data sets ####
# Read in the cleaned McKenna data # 
cleaned_mckenna_data <- read_csv(here::here("~/Desktop/Final Paper/Outputs/Data/cleaned_mckenna_data.csv"))
show_col_types = FALSE

# Read in the cleaned May data #
cleaned_may_data <- read_csv(here::here("~/Desktop/Final Paper/Outputs/Data/cleaned_may_data.csv"))
show_col_types = FALSE

# Read in the cleaned Collins data #
cleaned_collins_data <- read_csv(here::here("~/Desktop/Final Paper/Outputs/Data/cleaned_collins_data.csv"))
show_col_types = FALSE

#### Check that the raw McKenna data set is ready to be coded ####
# Code referenced from: https://tellingstorieswithdata.com/12-ijalm.html 
test_that("check class" , {
  expect_type(cleaned_mckenna_data$Text, "character")
  expect_type(cleaned_mckenna_data$Name, "character")
  expect_type(cleaned_mckenna_data$Username, "character")
  expect_type(cleaned_mckenna_data$Bio, "character")
  expect_type(cleaned_mckenna_data$Location, "character")
  expect_type(cleaned_mckenna_data$Followers, "double")
  expect_type(cleaned_mckenna_data$Following, "double")
  expect_type(cleaned_mckenna_data$URL, "character")
})

test_that("Check number of observations is correct", {
  expect_equal(nrow(cleaned_mckenna_data), 500)
})

#### Check that the raw May data set is ready to be coded ####
test_that("check class" , {
  expect_type(cleaned_may_data$Text, "character")
  expect_type(cleaned_may_data$Name, "character")
  expect_type(cleaned_may_data$Username, "character")
  expect_type(cleaned_may_data$Bio, "character")
  expect_type(cleaned_may_data$Location, "character")
  expect_type(cleaned_may_data$Followers, "double")
  expect_type(cleaned_may_data$Following, "double")
  expect_type(cleaned_may_data$URL, "character")
})

test_that("Check number of observations is correct", {
  expect_equal(nrow(cleaned_may_data), 500)
})

#### Check that the raw Collins data set is ready to be coded ####
test_that("check class" , {
  expect_type(cleaned_collins_data$Text, "character")
  expect_type(cleaned_collins_data$Name, "character")
  expect_type(cleaned_collins_data$Username, "character")
  expect_type(cleaned_collins_data$Bio, "character")
  expect_type(cleaned_collins_data$Location, "character")
  expect_type(cleaned_collins_data$Followers, "double")
  expect_type(cleaned_collins_data$Following, "double")
  expect_type(cleaned_collins_data$URL, "character")
})

test_that("Check number of observations is correct", {
  expect_equal(nrow(cleaned_collins_data), 13)
})
